extract folders from backend
