+++ massgen/backend/base.py
@@ -11,8 +11,8 @@ from enum import Enum
 from typing import Any, AsyncGenerator, Dict, List, Optional, Union
 
 from ..mcp_tools.hooks import FunctionHookManager, HookType
-from .utils.filesystem_manager import FilesystemManager, PathPermissionManagerHook
-from .utils.token_manager import TokenCostCalculator, TokenUsage
+from ..filesystem_manager import FilesystemManager, PathPermissionManagerHook
+from ..token_manager import TokenCostCalculator, TokenUsage
 
 
 class FilesystemSupport(Enum):
