+++ massgen/backend/chat_completions.py
@@ -29,8 +29,8 @@ from ..logger_config import log_backend_agent_message, log_stream_chunk, logger
 # Local imports
 from .base import FilesystemSupport, StreamChunk
 from .base_with_mcp import MCPBackend
-from .utils.api_params_handler import ChatCompletionsAPIParamsHandler
-from .utils.formatter import ChatCompletionsFormatter
+from ..api_params_handler import ChatCompletionsAPIParamsHandler
+from ..formatter import ChatCompletionsFormatter
 
 
 class ChatCompletionsBackend(MCPBackend):
