+++ massgen/filesystem_manager/_filesystem_manager.py
@@ -20,10 +20,10 @@ from datetime import datetime
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
-import massgen.backend.utils.filesystem_manager._workspace_tools_server as wc_module
+from . import _workspace_tools_server as wc_module
 
-from ....logger_config import get_log_session_dir, logger
-from ....mcp_tools.client import HookType
+from ..logger_config import get_log_session_dir, logger
+from ..mcp_tools.client import HookType
 from ._base import Permission
 from ._path_permission_manager import PathPermissionManager
