+++ massgen/formatter/_chat_completions_formatter.py
