+++ massgen/formatter/_claude_formatter.py
