+++ massgen/formatter/_response_formatter.py
